requirements.txt
Flask==1.0.2
Flask-Assets==0.12
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
Flask-Migrate==2.2.1
SQLAlchemy==1.3.5
mysqlclient==1.3.12
configobj==5.0.6
bcrypt==3.1.4
requests==2.20.0
python-ldap==3.1.0
pyotp==2.2.6
qrcode==6.0
dnspython==1.15.0
gunicorn==19.7.1
python3-saml
pyOpenSSL>=0.15
pytz>=2017.3
cssmin==0.2.0
jsmin==2.2.2
Authlib==0.10
Flask-SeaSurf==0.2.2
bravado-core==5.13.1
lima==0.5
pytest==5.0.1
pytimeparse==1.1.8
PyYAML==5.1.1
Flask-SSLify==0.1.5
